Print details of authentication workaround to terminal
Mitigates https://github.com/cli/cli/issues/297

This adds some more information to the output when authentication fails
due to `gh` being unable to open a browser. It communicates the details
of the workaround without the user having to search issues on the `gh`
repo to see how to get around it.