Allow for arbitrary OAuth hosts
Not all OAuth hosts use the same routes as GitHub, for example:
- Microsoft use /oauth2/v2.0/devicecode
- Google use /device/code
- Auth0 use /oauth/device/code

Similar differences are present for the authorise and access token
routes too.

This commit introduces a concept of a Server, which is a container for
the endpoints that the library uses. This is a replacement for Flow's
Hostname and as such is a breaking change.